[CI] Ensure notebook file is removed even if an exception occurs

+++ docs/notebook_validation.py
@@ -53,11 +53,13 @@ def execute(notebook_filename):
         print(
             f"Time taken for nbconvert : {elapsed:.2f} seconds for '{notebook_filename}'"
         )
-        os.remove(notebook_filename_out)
         return True
     except subprocess.CalledProcessError:
         print('Error executing the notebook "%s".\n\n' % notebook_filename)
         return False
+    finally:
+        if os.path.exists(notebook_filename_out):
+            os.remove(notebook_filename_out)
 
 
 def print_results(success, failed, skipped=[]):
